requirements.txt
pandas>=1.5.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # For PostgreSQL connection
PyYAML>=6.0.1
tqdm>=4.65.0
docker-compose>=1.29.2  # For managing Docker containers
streamlit>=1.24.0  # For interactive data apps
plotly>=5.15.0  # For interactive visualizations
numpy>=1.24.0  # For numerical computations
pycircos>=0.3.0  # For circular genome plots
matplotlib>=3.7.1  # For plotting
biopython>=1.81  # For genomic data handling
